Styling on Interview Setup
