+++ src/styles/_FoundStudent.scss
@@ -1,0 +1,7 @@
+.found-student {
+  background: $card-bg;
+  padding: 1rem;
+  margin-top: 2rem;
+  border-radius: 1rem;
+}
+
