+++ src/styles/main.scss
@@ -11,12 +11,12 @@
 @import "InterviewEnd";
 @import "LoginForm";
 @import "Question";
-@import "SelectedStudentModal";
-@import "SelectStudentContainer";
+@import "InterviewSetup";
+@import "StudentSearch";
+@import "FoundStudent";
 @import "StudentInterview";
 @import "StudentInterviewContainer";
 @import "StudentInterviewModal";
-@import "StudentProgramSearchForm";
 @import "StudentQuestionCard";
 @import "StudentQuestionContainer";
 @import "StudentQuestionModal";
